+++ src/IEntry.ts
@@ -1,0 +1,4 @@
+export interface IEntry<K, V> {
+  key: K,
+  value: V,
+}